Fix coalescence (#170)
Currently it did not sort correctly, which could lead to errors like two smaller fleets combining to be the biggest fleet.

Also fix the schema to be reflective of the correct data types.

+++ kaggle_environments/envs/kore_fleets/starter_bots/java/test/KoreJsonTest.java
@@ -38,7 +38,7 @@ public class KoreJsonTest {
         Path observation = Paths.get("bin", "test", "observation.json");
         String raw = Files.readString(observation);        
 
-        float[] kore = KoreJson.getDoubleArrFromJson(raw, "kore");
+        double[] kore = KoreJson.getDoubleArrFromJson(raw, "kore");
         Assert.assertEquals(kore[3], 1.372, 0.0001);
     }
